+++ packages/to-json-schema/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v1.2.0 (May 17, 2025)
 
 - Add support for title, description and examples in `metadata` action (pull request #1189)
 - Add new override configurations to override default behaviour of JSON Schema conversion (pull request #1197)